scripts/requirements.txt
accelerate
certifi
cffi
charset-normalizer
colorama
contourpy
cycler
filelock
fonttools
fsspec
hf-xet
huggingface-hub
idna
importlib-resources
inquirerpy
jinja2
joblib
kiwisolver
markdown-it-py
markupsafe
matplotlib
mdurl
mpmath
networkx
numpy
nvidia-cublas-cu12
nvidia-cuda-cupti-cu12
nvidia-cuda-nvrtc-cu12
nvidia-cuda-runtime-cu12
nvidia-cudnn-cu12
nvidia-cufft-cu12
nvidia-cufile-cu12
nvidia-curand-cu12
nvidia-cusolver-cu12
nvidia-cusparse-cu12
nvidia-cusparselt-cu12
nvidia-nccl-cu12
nvidia-nvjitlink-cu12
nvidia-nvtx-cu12
packaging
pandas
pfzy
pillow
prompt-toolkit
psutil
pycparser
pygments
pyparsing
python-dateutil
pytz
pyyaml
regex
requests
rich
rich-argparse
safetensors
scikit-learn
scipy
setuptools
six
sympy
threadpoolctl
tokenizers
torch
torchvision
tqdm
transformers
triton
typing-extensions
tzdata
urllib3
wcwidth
zipp
